+++ LightBlue.WorkerService/WorkerHostSettings.cs
@@ -1,0 +1,10 @@
+﻿namespace LightBlue.WorkerService
+{
+    public class WorkerHostSettings
+    {
+        public string Assembly { get; set; }
+        public string RoleName { get; set; }
+        public string ServiceTitle { get; set; }
+        public string Configuration { get; set; }
+    }
+}